TranslationController层大小写接口完成
