+++ mubang/src/main/java/com/lzj/demo/service/TransactionManagementConfiguration.java
@@ -1,0 +1,28 @@
+package com.lzj.demo.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.datasource.DataSourceTransactionManager;
+import org.springframework.transaction.TransactionManager;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+import org.springframework.transaction.annotation.TransactionManagementConfigurer;
+
+import javax.sql.DataSource;
+
+/**
+ * @author 林子键
+ * @version 1.0
+ */
+@Configuration
+@EnableTransactionManagement
+public class TransactionManagementConfiguration implements TransactionManagementConfigurer {
+
+    @Autowired
+    private DataSource dataSource;
+
+    @Override
+    public TransactionManager annotationDrivenTransactionManager() {
+        return new DataSourceTransactionManager(dataSource);
+    }
+
+}
